+++ CHANGELOG.md
@@ -7,7 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
--- Nothing yet
+- Update to JPAN 0.7.0 and improve SCMP error handling.
+  [#18](https://github.com/netsec-ethz/jpan-cli/pull/18)
 
 ## [0.2.2] - 2026-03-27